src/datascience/components/data_transformation.py
import sys
import numpy as np
import pandas as pd
from dataclasses import dataclass
import os
from sklearn.preprocessing import OneHotEncoder,StandardScaler
from sklearn.pipeline import Pipeline
from  sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from datascience.exception_handling import CustomException
from datascience.logger import logging
