requirement.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
catboost
xgboost
pymysql
python-dotenv
sqlalchemy
mysql-connector-python


# -e .